Potential fix for NPE when initialising daycare
The NPE would occur when the DaycareGenerations.java  calls the getDependency function, which then tries to get to the (impactor) plugin instance for the manager, but if said plugin isnt fully initialised yet, it will cause the daycare to throw a NPE and therefore not work.

The attempt was actually shamelessly stolen from the GTS repo, which also uses the dependcymanager, but instead of grabbing it from another plugin, it creates its own instance, which then works out fine. My attempt here was to do the same with the daycare and therefore preventing it from crashing onInit (example stacktrace: https://mcpaste.io/5d3c035055ef1616).

PS: The code is untested, but in theory, this is how it should work.